refactor: folder, shared 페이지 수정
