+++ pages/shared/index.tsx
@@ -1,49 +1,43 @@
 import Layout from "@/components/layout/layout/Layout";
-import { useFolder } from "../api/useFolder";
-import FolderInfo from "@/components/folder-info/FolderInfo";
-import SearchBar from "@/components/search-bar/SearchBar";
-import CardList from "@/components/card-list/CardList";
-import Card from "@/components/card/Card";
+import FolderInfo from "@/components/shared/folder-info/FolderInfo";
+import SearchBar from "@/components/shared/search-bar/SearchBar";
+import CardList from "@/components/ui/card-list/CardList";
+import Card from "@/components/ui/card/Card";
 import formatDate from "@/lib/formatDate";
 import { getElapsedTime } from "@/lib/getElapsedTime";
 import styles from "./Shared.module.scss";
 import classNames from "classnames/bind";
+import { useFolder } from "@/pages/hooks-test/useFolder";
 
 const cx = classNames.bind(styles);
 
-interface CardProps {
-  id: number;
-  url: string;
-  imageSource: string;
-  alt: string;
-  elapsedTime: string;
-  description: string;
-  createdAt: string;
-}
-
 export default function Shared() {
-  const { data } = useFolder();
+  const { data, isLoading, isError } = useFolder();
+
+  if (isLoading) return <div>loading</div>;
+  if (isError) return <div>error</div>;
+  if (!data) return <div>data가 없음</div>;
 
   return (
     <Layout isSticky={false}>
       <div className={cx("container")}>
         <FolderInfo
-          folderName={data?.folder.name}
-          ownerName={data?.folder.owner.name}
-          profileImage={data?.folder.owner.profileImageSource}
+          folderName={data.name}
+          ownerName={data.owner.name}
+          profileImage={data.owner.profileImageSource}
         />
         <div className={cx("item")}>
           <SearchBar />
           <CardList>
-            {data?.folder.links.map((item: CardProps) => (
+            {data.links.map((link) => (
               <Card
-                key={item.id}
-                href={item.url}
-                imageSource={item.imageSource}
+                key={link.id}
+                href={link.url}
+                imageSource={link.imageSource}
                 alt={"card"}
-                elapsedTime={getElapsedTime(item.createdAt)}
-                description={item.description}
-                createdAt={formatDate(item.createdAt)}
+                elapsedTime={getElapsedTime(link.createdAt)}
+                description={link.description}
+                createdAt={formatDate(link.createdAt)}
               />
             ))}
           </CardList>
